EMPLOYEE INFORMATION SYSTEM
This is a simple project entitled Employee Information System. This is a web-based application developed in Python using Django Framework. This application manages the companies employee information. I used different libraries on this website to give the end-users a pleasant user interface and a better experience while using the application. This can help a certain company to easily retrieve, store, and manage their employees' information